requirements.txt
streamlit
pyyaml
anthropic
matplotlib
